test: Migrate proxy to containers
