+++ .gitignore
@@ -9,7 +9,6 @@ yarn.lock
 google-generated-credentials.json
 _START_PACKAGE
 .env
-.services.json
 .vscode/*
 !.claude
 !.vscode/extensions.json
