+++ packages/testing/containers/package.json
@@ -19,7 +19,7 @@
     "stack:clean:networks": "docker network ls --filter 'label=org.testcontainers=true' -q | xargs -r docker network rm 2>/dev/null",
     "stack:clean:all": "pnpm run stack:clean:containers && pnpm run stack:clean:networks",
     "services": "tsx ./n8n-start-stack.ts --services-only",
-    "services:clean": "docker ps -aq --filter 'name=n8n-svc-*' | xargs docker rm -f 2>/dev/null; rm -f ../../../packages/cli/bin/.env ../playwright/.services.json",
+    "services:clean": "docker ps -aq --filter 'name=n8n-svc-*' | xargs docker rm -f 2>/dev/null; rm -f ../../../packages/cli/bin/.env",
     "lint": "eslint . --quiet",
     "lint:fix": "eslint . --fix"
   },
@@ -33,6 +33,7 @@
     "@testcontainers/postgresql": "^11.0.3",
     "@testcontainers/redis": "^11.0.3",
     "get-port": "^7.1.0",
+    "mockserver-client": "^5.15.0",
     "kafkajs": "catalog:",
     "testcontainers": "^11.11.0"
   }
