+++ packages/testing/playwright/tests/e2e/ai/hitl-for-tools.spec.ts
@@ -52,9 +52,9 @@ const hitlForToolsTestConfig = {
 
 test.use(hitlForToolsTestConfig);
 test.describe('HITL for Tools @capability:proxy', () => {
-	test.beforeEach(async ({ n8n, proxyServer }) => {
-		await proxyServer.clearAllExpectations();
-		await proxyServer.loadExpectations('hitl-for-tools');
+	test.beforeEach(async ({ n8n, services }) => {
+		await services.proxy.clearAllExpectations();
+		await services.proxy.loadExpectations('hitl-for-tools');
 		await n8n.canvas.openNewWorkflow();
 	});
