+++ packages/testing/playwright/tests/e2e/ai/langchain-agents.spec.ts
@@ -77,9 +77,9 @@ async function setupBasicAgentWorkflow(n8n: n8nPage, additionalNodes: string[] =
 
 test.use({ capability: 'proxy' });
 test.describe('Langchain Integration @capability:proxy', () => {
-	test.beforeEach(async ({ n8n, proxyServer }) => {
-		await proxyServer.clearAllExpectations();
-		await proxyServer.loadExpectations('langchain');
+	test.beforeEach(async ({ n8n, services }) => {
+		await services.proxy.clearAllExpectations();
+		await services.proxy.loadExpectations('langchain');
 		await n8n.canvas.openNewWorkflow();
 	});
