+++ packages/testing/playwright/tests/e2e/ai/langchain-chains.spec.ts
@@ -40,9 +40,9 @@ async function executeChatAndWaitForResponse(n8n: n8nPage, message: string) {
 
 test.use({ capability: 'proxy' });
 test.describe('Langchain Integration @capability:proxy', () => {
-	test.beforeEach(async ({ n8n, proxyServer }) => {
-		await proxyServer.clearAllExpectations();
-		await proxyServer.loadExpectations('langchain');
+	test.beforeEach(async ({ n8n, services }) => {
+		await services.proxy.clearAllExpectations();
+		await services.proxy.loadExpectations('langchain');
 		await n8n.canvas.openNewWorkflow();
 	});
