+++ packages/testing/playwright/tests/e2e/ai/langchain-memory.spec.ts
@@ -50,9 +50,9 @@ async function verifyChatMessages(n8n: n8nPage, expectedCount: number, inputMess
 
 test.use({ capability: 'proxy' });
 test.describe('Langchain Integration @capability:proxy', () => {
-	test.beforeEach(async ({ n8n, proxyServer }) => {
-		await proxyServer.clearAllExpectations();
-		await proxyServer.loadExpectations('langchain');
+	test.beforeEach(async ({ n8n, services }) => {
+		await services.proxy.clearAllExpectations();
+		await services.proxy.loadExpectations('langchain');
 		await n8n.canvas.openNewWorkflow();
 	});
