+++ packages/testing/playwright/tests/e2e/ai/langchain-vectorstores.spec.ts
@@ -10,9 +10,9 @@ async function waitForWorkflowSuccess(n8n: n8nPage, timeout = 10000) {
 
 test.use({ capability: 'proxy' });
 test.describe('Langchain Integration @capability:proxy', () => {
-	test.beforeEach(async ({ n8n, proxyServer }) => {
-		await proxyServer.clearAllExpectations();
-		await proxyServer.loadExpectations('langchain');
+	test.beforeEach(async ({ n8n, services }) => {
+		await services.proxy.clearAllExpectations();
+		await services.proxy.loadExpectations('langchain');
 		await n8n.canvas.openNewWorkflow();
 	});
