+++ packages/testing/playwright/tests/e2e/ai/workflow-builder.spec.ts
@@ -31,10 +31,10 @@ test.use({
 });
 
 test.describe('Workflow Builder @auth:owner @ai @capability:proxy', () => {
-	test.beforeEach(async ({ setupRequirements, proxyServer }) => {
+	test.beforeEach(async ({ setupRequirements, services }) => {
 		await setupRequirements(workflowBuilderEnabledRequirements);
-		await proxyServer.clearAllExpectations();
-		await proxyServer.loadExpectations('workflow-builder');
+		await services.proxy.clearAllExpectations();
+		await services.proxy.loadExpectations('workflow-builder');
 	});
 
 	test('should show Build with AI button on empty canvas', async ({ n8n }) => {
