+++ packages/testing/playwright/tests/e2e/chat-hub/fixtures.ts
@@ -37,16 +37,16 @@ export const test = base.extend<ChatHubFixtures>({
 	},
 
 	chatHubProxySetup: [
-		async ({ proxyServer }, use) => {
+		async ({ services }, use) => {
 			// Setup
-			await proxyServer.clearAllExpectations();
-			await proxyServer.loadExpectations('chat-hub', { strictBodyMatching: true });
+			await services.proxy.clearAllExpectations();
+			await services.proxy.loadExpectations('chat-hub', { strictBodyMatching: true });
 
 			await use(undefined);
 
 			// Teardown
 			if (!process.env.CI) {
-				await proxyServer.recordExpectations('chat-hub', {
+				await services.proxy.recordExpectations('chat-hub', {
 					dedupe: true,
 					transform: (expectation) => {
 						const response = expectation.httpResponse as {
